fix: #2512 notification for aml checklist
