+++ packages/aml-checklist/src/components/hocs/__tests__/error-boundary.tsx
@@ -1,52 +1,35 @@
 import * as React from 'react'
 import { shallow, mount } from 'enzyme'
-import toJson from 'enzyme-to-json'
 import { ErrorBoundary, ErrorState } from '../error-boundary'
-import errorMessages from '../../../constants/error-messages'
-import { ErrorData } from '../../../reducers/error'
 
 jest.mock('../../../utils/route-dispatcher')
-
-const Children = () => <div>I am a component!</div>
-const props = {
-  children: Children,
-  errorThrownComponent: jest.fn(),
-  componentError: {
-    type: 'COMPONENT',
-    message: errorMessages.DEFAULT_COMPONENT_ERROR,
-  } as ErrorData,
-}
+jest.mock('rc-animate')
 
 describe('ErrorBoundary', () => {
   it('should match a snapshot when no error', () => {
-    expect(toJson(shallow(<ErrorBoundary {...props} />))).toMatchSnapshot()
+    expect(shallow(<ErrorBoundary />)).toMatchSnapshot()
   })
 
   it('should match a snapshot when has an error', () => {
-    const component = shallow(<ErrorBoundary {...props} />)
+    const component = shallow(<ErrorBoundary />)
     component.setState({
       hasFailed: true,
     })
-    expect(toJson(component)).toMatchSnapshot()
+    expect(component).toMatchSnapshot()
   })
 
-  it('should call the errorThrownComponent and sets the state to hasFailed when it catches', () => {
+  it('should call the notification sets the state to hasFailed when it catches', () => {
     ;(console.error as any) = jest.fn()
-    const DangerousChild = (props: { someProp?: false }) => {
-      if (!props.someProp) {
-        throw new Error('Catch me if you can')
-      }
-      return <div />
+    const DangerousChild = () => {
+      throw new Error('Catch me if you can')
     }
-    const newPops = { ...props, children: <DangerousChild /> }
 
-    const component = mount(<ErrorBoundary {...newPops} />)
+    const component = mount(
+      <ErrorBoundary>
+        <DangerousChild />
+      </ErrorBoundary>,
+    )
     expect(DangerousChild).toThrow()
-    expect(newPops.errorThrownComponent).toHaveBeenCalledTimes(1)
-    expect(newPops.errorThrownComponent).toHaveBeenCalledWith({
-      type: 'COMPONENT',
-      message: errorMessages.DEFAULT_COMPONENT_ERROR,
-    })
     expect((component.state() as ErrorState).hasFailed).toBe(true)
   })
