+++ packages/aml-checklist/src/components/pages/client-search/index.ts
@@ -1,2 +1,3 @@
 import ClientSearch from './client-search'
+
 export default ClientSearch
