--- packages/aml-checklist/src/constants/auth.ts
@@ -1,5 +1,0 @@
-import { StringMap } from '@/types/core'
-
-export const LOGIN_TYPE = {
-  CLIENT: 'CLIENT',
-} as StringMap
