--- packages/aml-checklist/src/constants/session.ts
@@ -1,1 +1,0 @@
-export const LOCAL_STORAGE_SESSION_KEY = 'reapit-app-store-session'
