+++ packages/aml-checklist/src/types/core.ts
@@ -1,6 +1,5 @@
 import Routes from '../constants/routes'
 import ActionTypes from '../constants/action-types'
-import { ErrorState } from '../reducers/error'
 import { ChecklistDetailState } from '@/reducers/checklist-detail'
 import { ResultState } from '@/reducers/result'
 import { IdentityTypesState } from '@/reducers/identity-types'
@@ -37,7 +36,6 @@ export interface FetcherParams<T> {
 }
 
 export interface ReduxState {
-  error: ErrorState
   checklistDetail: ChecklistDetailState
   result: ResultState
   identityTypes: IdentityTypesState
